requirements.txt
Flask==3.0.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.126.0
requests==2.31.0
openai==1.12.0
python-dateutil==2.8.2